Bugs, refactoring, fixes and better
The final changes to make the new approach to defining tables is close
to done. I found some mistakes, I found some forgotten tables. I
refactored some code, moved things from DatabaseSchemaHelper, etc.

But I can now actually copy a pgsql database to sqlite and it works.
There are still bugs to be found, but the majority of it seems to be
working.

Since I redefined the Brewtarget::DBTypes enum, I had some fixing to do
in the OptionDialog. I had made the mistake of assuming those values
would be the same as the index of the options. Not a hard fix, but
annoying.

I bumped the DB version to remove the two mystery columns from
brewnotes.

Finally, I removed some random qDebug statements that I had probably
committed earlier.