Signal on load and inventory fixes
The first issue was the result of how we load the data now. When a
recipe is loaded, there are four(ish) fields we calculate but only when
recipe->recalcall() is called. The new caching method doesn't fire the
right signals, so this was never called and weird things happened. I
modified the process so that loading a reciping in MainWindow causes
recalcAll to be called. This is probably over kill.

The second problem was with the inventory tables, specifically the yeast
table. Because it uses a different name for the amount (quanta), bad
things happened. I've fixed it, but I really want to rework the entire
inventory table system. What we have now is kinda complex and hard to
understand. I think I can reduce it all down to one table and some
magix.

Anything else in this commit is remove qDebug statements.