setting name needs to honor cacheOnly
The editors needed to honor the cacheOnly flag when setting the name.
Since the cacheOnly flag is in the object (eg, hop) and not the parent
(BeerXMLElement), I have to pass that in. Otherwise, it throws some
uncomfortable errors about updating id -1.

Maybe I should put the cacheOnly stuff into the parent object. It would
save some code and things.