We write strings, not enums
Imagine my surprise when I realized we right the untranslated strings
for things like mash step types, hop use, etc. to the database instead
of the enumerated value.

This fixes many of the issues introduced by that oversight.

And we need to fix this properly. The enumerated value should be
written to the database, not a string. String comparisons are just not
good.