removing static column names
This is the part I was dreading. Now that I have declared we cannot know
the property names, I have to crawl through Database and make sure we
are using dynamic references instead of hard coded string names. It has
resulted in gems like:
  QString("select %2.%1 from %2, %3, %4 where %3.%6 = %5 and %3.%6 = %4.%7 and %4.%8 = %2.%9")

In return, the XML output code has gotten a lot nicer. I was able to
replace almost all of the boilerplate with a nice loop. I think it will
work.

The code compiles. I am sure it dumps core in a thousand ways. I will
fix those later.