Removed all propertyToColumn getting keys
The problem with work like this is the standards are very fluid, and
sometimes you have to stop and backfill. There were a number of cases
where I was using propertyToColumn() to get foreign key names. That
doesn't work anymore, so I had to convert all of those to making the
proper calls.

I sorted out getInventoryAmt as well. Which was horrific and I really
want to rethink how we are doing inventory.

The code compiles and runs, but there is a problem linking mashes and brewnotes to
their recipes. That will be my next major bug hunt.