Wasn't loading hops, fermentables, etc. properly
Chasing lots of smallish bugs with large effects. Used the wrong index
name for doing the initial loads (childIndexName() instead of the
correct inRecIndexName()). Once that was fixed, lots of things are
working better.

I am trying to cache the inventory amounts now. This is proving
difficult. I think I need to set it up such that the parent signals it's
children that the inventory amount has changed, thei children update
their cache and signal that change.

There may be a third path here. The children don't store the inventory
amount, but get it from the parent. I would need a reliable method to
know a thing is a child (deleted = false, display = false?) and to find
the parent. That gets sloppy fast when a parent is (say) deleted.
We never actually delete though, so that may have value.