Inventory, made better with signals!
I had been trying to avoid this, but the last massive number of
redundant queries were for getting and setting inventory. The queries
themselves were also pretty gross. So I made them grosser, but way faster.

The first thing that happened was each of the four tables (fermentables,
hops, misc and yeast) got a new column inventory_id. Since we have a
many-to-one relationship between ingredients and inventory, this seemed
a good fit.

The hard part was populating that correctly. If you value your sanity, I
would suggest not looking at migration_helper_8. It adds the new column,
deletes any inventory table row that is linked to a child, updates all
of the parents to point to the right rows, adds an inventory row for any
parent that doesn't already have one (needed, believe me) and finally
updates all the children to have the same value as the parent.

When a new ingredient is added, we just blank copy the parent's
inventory_id. Easy peasy.

The next thing that happened is each of those four objects got two new
attributes to cache the inventory amount and the inventory id (which is
needed in later steps). The TableSchemas were updated to include the new
inventory_id attribute where required.

Database::setInventory now issues a new signal when an inventory is
updated. The signal includes the table that was updated, the
inventory_id updated and the new value. I may fine tune this signal
later, because it is pretty gross right now.

Each of the TableModels now has a new slot to catch that signal. If the
signal has the right Brewtarget::DBTable and the inventory_id matches
one of it's observed elements, then we update the cached value (but do
NOT write it to database -- it's already there).

This required a lot of fussing around in Database and BeerXMLElement to
get the call signatures right, etc.

What's odd is it seems to work. Along the way, I found a few bugs and
cleaned those up. I also de-linted things along the way.

I need to go over database.cpp and clean out the dead code again. And
then I turn the blow torch to DatabaseSchemaHelper.