Instruction bugs, more cleaning deprecated code
Some of the things that I saw as deprecated are only deprecated in QT
5.13 and the new methods don't exist prior to that. I had to do some #if
work to make everything happy.

The main part of this commit is cleaning up the Instructions. I had
not remembered to add them to the database when creating one, so I fixed
that. I added a new parameter to the TableSchema so you can name one
attribute to have the increment/decrement triggers added. It's over
kill, but makes it easier should we ever want to (say) do the mashsteps
to use it.

I also had to fix a few smaller queries to get the instructions mostly
working. Attempting to re-order them via the GUI isn't working, and I
need to figure out why. But this is enough for one day.

I started cleaning up/out DatabaseSchemaHelper. All of the migrate
methods now use DatabaseSchema and TableSchema. I need to figure out
which of the thousands of methods I can clean out next.

I added a version number (config_version) to the config file. This
should make it easier to update as we need to in the future. This caused
me to re-arrange the readPersistentSettings method a little.