Instructions work, inventory fixes continue
Figured out all the problems with the instructions. They work now,
including being able to change the order.

Adding a new item now does the right thing with inventory. That ended up
a little different than I had expected. The inventory amount actually
gets set in the editor, after the base object is created and written to
the db.

Discovered that I hadn't created the setInventoryId methods for anything
other than fermentables, so that got fixed.

Updated the docs a little too.