DatabaseSchemaHelper is clean
Oh. That felt good. Righteous. All of the methods we don't use, all the
constants, all of the hard to read was removed from
DatabaseSchemaHelper.

It compiles and runs, so that's good enough for a commmit. I will test a
few things just to make sure I haven't missed things.

Righteous.