Inventory again, and what have I done
So sqlite and postgresql do not like each other's SQL. The upgrade from
7 -> 8 worked for sqlite, then worked for postgresql and now works
for both.

The inventory tables continue to plague me. I realized that adding the
inventory_id to the ingredient tables and leaving the [ingredient]_id
(eg, hop_id) column in the inventory table left me with a circular
dependency -- ingredients pointed to inventory pointed to ingredient. So
I dropped the foreign key from inventory to ingredien.

That required I change many things in how the ingredients interact with
inventory. Mostly, ingredients needed to provide their inventory_id when
updating inventory, which required changes in BeerXMLElement and
database.

I think I am close to doing the merge request.