bt_alltables gone, creating postgresql db works
Some fairly minor tweaks this time. I've removed the bt_alltables and
its dependencies as it replaced by TableSchema and DatabaseSchema.

A small fix to the options dialog so it won't erase some settings.

Duplicating a sqlite database to postgres works now, along with properly
resetting the sequences.

Changing the relationship between ingredient and inventory requires the
inventory tables to be loaded first, so that the constraint on the
ingredient table is satisfied. That took some thinking, but I think I
found a good solution.