Signals and other small fixes
The new inventory signal wasn't getting properly set or caught by one of
the table models.

Reformatted some of mainwindow so I could understand the code.

Found a small bug in MashWizard. The final mash calculations were
ignoring the lauter tun deadspace. I probably should test the nosparge,
but there's not really a lauter tun in nosparge.

Updated the documentation and added some comments.

With my reworking of the inventory system, the invIndexName() method
doesn't make sense. It has been removed.

The reworking of the inventory system also required getInventory() to be
reworked. The query is rather ugly. I may find a way to simplify it
later.