I am about to break it all
The BeerXML for water makes almost no sense to me. It is great for
showing the anions/cations you want, but I don't see an easy way to get
from that to the actual salt additions you want to make. So if I know I
need 120 ppm Ca, did I get there using CaCl2, CaSO4 or CaCO3?

I supposed I could do this clever like, and an addition of CaCl2 would
have the proper values for Ca and Cl and nothing else. That seems wrong
to me.

So I am about to completely redefine the tables. It means I am breaking
a BeerXML defined table (water). I will have to get clever in the toXml
routines.

This branch (and any database I am pointing to) will be broken for a
while. I want this in place because I actually have a quasi-working
system that I may need to revert to repeatedly as I try to figure this
out.