A weird merge, and I missed misc
I really need to fix how we record somethings in the database. I dislike
storing things like "use" as strings. I caught this problem earlier with
hops and yeast, but waited until this commit to fix misc.

There was also a weird hang over from the previous git merge that was
found when fixing this bug.