Misc tree (mostly) works, drag/drop kind of works
The misc tree is display what I want, although I need to figure out how
to auto-expand the tree. It is a setting somewhere, but I don't remember
what it is.

Drag/drop works, but the Recipe object doesn't like a QList<Salt*> as a
metatype. I think I've done everything I need to do but it still throws
the error.

I need to work on the table tomorrow.