Kinda sorta working
Ok. This roughs in the basics. Dropping a salt onto the table and adding
an amount causes many of the right things to happen. It is being done
dynamically, so nothing is saved yet.

Many, many things are missing. Like pH calculations, like adding the
proper salts to the Misc table, like deleting a salt you've added, etc.
But this a basic proof of concept or rought draft.

I think the re-use of the BtTree is clever, but I am not sure it
actually buys me anything. I am going to try a different approach that
doesn't require as many signal/slots and isn't quite so clever.