Much better idea
Ok. This is working way better for me.

I reworked the main screen and I think it makes more sense now. I added
a new "addedTo" state called ratio and moved 'both' to 'equal'. I
removed the 'magic' button, and need to clean out the evenSplit field. I
think the input screens make more sense.

I swapped out the lineEdit for the mash and sparge RO percents for
spinboxes. It works much better, I think. I've also plugged all of that
math in.

A salt marked as 'ratio' will be added to the mash as entered,
and added proportionally to the sparge -- so if you are adding 8g of
CaSO4 'ratio' and the sparge is 1/2 the mash, then you will add 8g to
the mash water and 4g to the sparge. Equal would mean adding 8g to both.

I am about 70% done. I need to do the pH stuff; I need to save the
information and load it from the db; I need to figure out how I'm lying
to BeerXML about what I've done.