Salt table changes for acids
I am really trying to avoid recreating the Misc table, but I am not
succeding. I simply have no way of knowing what things in the misc table
are what I'm interested in, and I have no way of figuring it out. So I
am forced to take the pieces I need from misc and re-implement them.

This commit extends the salt table to be able to do acid. I added three
columns to the table: is_acid, amount_is_weight and percent_acid. I've
built in a large number of assumptions (lactic and phosphoric acid are
always liquids, etc) that I will have loads of fun untangling later.

Since we are treating these as salts, I have to worry about mash and
sparge additions. I added totalAcidWeight() to saltTableModel to do the
right maths. It returns the actual acid weight, because anything else
was confusing.

The maths look about right. I am a little concerned that I am showing a
pH of 5.35 and Kai's spreadsheet is saying 5.31. I am uncertain how
concerned I should be.

I need to bend some attention to the layout of the page. I think I can
compress the space a little by moving the totals section to the left,
and putting the sliders in a 3x2 matrix that spans the entire bottom of
the window. I would like to figure out how to make the table expand to
the space it's in.