More small issues
Well, the bugs are small but the effects are noticeable.

I found out that the default-db.sqlite is reporting a 2.1.0 schema
version, aka, version 3. We are on version 8 now, and I don't think
that's going to be helpful.

In the process of upgrading that, I found three issues.

The upgrade from v3 -> v4 was messier than it had to be and broke. I
replaced a lot of work with an alter table rename.

Originally, the upgrade from v5 -> v6 built the bt_alltables table. That
table gets removed in v8, and so I had just gutted the original code.
This broke the v8 upgrade because it was trying to drop a table that
didn't exist. It was fixed by adding " IF EXISTS" in the v8 drop.

There was a small typo in Database::populateChildTablesByName where I
said '4%' instead of '%4'. It made a difference to QString, and things
broke.

I added a logI method to Brewtarget which makes for nicer messages in
places.

Finally, I am really bad at pulling out my qDebug statements. Hopefully,
I found the last of them. I wouldn't hold my breath though.