So close, yet so far.
So my last task was to link the Misc entry to the appropriate salt. In
poking at the problem, it has occurred to me I've solved the data
storage problem all wrong.

This is now a dead branch. I will use it for cherry picking, but that's
about it. Murphy weeps.