More bugs?
This is an odd edge case where we create teh database from scratch.
Since the DB is created from the most recent definitions, we don't need
to run the upgrade process. We do need to shove the proper values into
the settings table.

Also added the dbType() parameter to the call from database to
databaseschemahelper::create because the default wasn't doing the right
thing.