Others
So it would seem that the original idea with the "other" methods (eg,
Hop::Hop(other)) did a lazy load -- the individual attributes would only
be populated when the getter was called. The only thing that was copied
were the cached values.

I broke that model. If you used this method before this fix, you would
only get those values which were cached in the BeerXMLElement. I had to
properly copy those values from the other to the new, which is what this
patch does.