misc uses amount, not amount_kg
The Q_PROPERTY for misc said 'amount', but my TableSchema config defined
it as kpropAmountKg, which was wrong. I fixed TableSchema to define the
table properly, I fixed Misc to save the right property and I had to
make a small fix in recipe to access the right property.

Someday, I will get rid of all the amount_kg, amount_l, etc. But not
today.