Mash fixes
I had the wrong table for creating a new mash. It caused lots of
problems.

MashDesigner wasn't adding the newly created mash to the recipe. You
could design all the mashes you wanted, but nothing got saved.

MashEditor returned if mashObs was null. Which it would be if you were
editing a brand new, never before seen mash. I fixed that to create the
Mash object, do the insert and add the mash to the recipe. This required
me to get ride of setEquipment and add setRecipe to MashEditor, and to
change all the calls to that method.