Closes #457 - text notes on Brew Days gone
A simple bug with an unfortunate side effect.

In defining the TableProperty for brewnotes, I didn't include the notes
column. I am putting it back.

The problem is that the upgrade from version 7 -> 8 drops two columns
from brewnote. In sqlite, this means you create a temp table, copy the
data from the old table to the temp, drop the old table and rename the
temp.

Because I didn't define the notes column, it was dropped as well.
Anybody impacted will either need to find an existing backup that is
still version 7 and copy that into place, or will need to manually add
the column to the brewnote table. There is some indication that the data
was preserved, even though the column was dropped?