pH is responding correctly
The pH field is no displaying enough information to be useful.

As a bonus, I found and fixed an issue that's driven me nuts for a
while. If you enter "ml" instead of "mL", the right thing happens now.
Both are considered correct, although mL is preferred to avoid
confusion.