BtDigitWidget now does unit and scale
At least for masses and generics. So far, that's all I've needed and I
am not sure I want to add more than I need.

Most of the other changes are to make that work.

I've made some pleasing changes to WaterDialog to be more loopy. This
required a few new methods in salt and water, as well as some new
constants.

Other small bug fixes (like learning that == != !=) are included for
free.