Better messages and nosparge
@dpettersson suggested I should have a better error message than my
usual sarcastic debug statements.

If you try to open the water dialog without a recipe selected or a mash
defined for that recipe, you will get a popup error saying you need to
define a mash first. I also moved the call to WaterDialog::observeRecipe
until as late as possible.

It also occurred to me I was not doing the right thing for nosparge
mashes, so I fixed that. If you don't have a sparge step in your mash,
then the "Mash", "Ratio" and "Equal" options aren't enabled.