Breaking up is hard to do
database.cpp is about 5100 lines long. I've been itching to take it
apart for a long time now. This is my first swing at doing so.

First, I renamed BeerXMLElement to Ingredient. I think that expresses
what they are a lot better and freed the name up that I wanted to use.

I then created the BeerXML class, which does all things related to
translating to and from XML. I think this will become important as
BeerJSON is being actively worked on and we should support that too.

That took 1600 lines from database.cpp.

My next pass will be to take all of the initialization code into its own
class. That should leave the database class doing one thing --
interfacing with the database.