Closes #507 - adding equipment to new recipe
My usual process in creating a new recipe is to create the recipe; set
the style and equipment; add fermentables, hops and yeast; and then
define the mash.

The undo code missed this possible path and assumed there would always
be a mash.

I simply moved the creating the undo objects to earlier in the code,
where we had already checked to make sure we had a mash. It had the
happy side effect of also making sure the undo actions were pushed
regardless of how the user answered the dialog, just like modifying the
properties was.