Closes #532 -- Problem with amounts in inventory
The code was working as designed, I just had a bad design. I had missed
this use-case, and so the original's inventory_id was copied along
with the rest.

The fix was to add a boolean to the new* calls in Database that allowed
us to specify if we wanted a new inventory row or not. Since this
allowed me to declutter some of the code, I think it is a reasonable
approach.

The only path I could find that needed this was from
BtTreeModel::copySelected, so that method got modified to request the
new inventory row. I had considered putting the code to add a new
inventory row here, which had the benefit of keeping the complexity
close to where it was used. But I didn't like the model having to open a
transaction, so I put into the Database methods.

My quick testing makes me think this is working better.