Cleaning constructors
I have cleaned up the constructors when loading an object from the
database. I messed up my design goals, and the constructors used the
column names directly. The correct thing to do was go through
TableSchema to translate the property names to column names. This had
some knock on effects.

I was able to remove an unsued constructor from each primitive. I also
was able to get rid of a newNamedEntity template that was only used in
two places.

NamedEntity's constructors also got a make over. At the price of reading
the database immediately after creating a record, I was able to simplify
the signature and I think it reads much better.

I have tested this by creating a new primitive, saving and reloading it.
At this point, everything by the mash and/or mashsteps is behaving.