Closes #564 -- updating the database is misbehaving
This was several problems all sort of rolled into one.

The initial problem was likely me being clever. Instead of updating
existing things and inserting new things, we updated existing things and
created new things and then updated them. This had a cascade failure
that ended up showing blank entries in the trees. They were properly
loaded on restarting brewtarget, but the idea was wrong.

The fix was to update that which existed, and create that which didn't.
This required a lot of reworking in TableSchema. The m_properties are
keyed by the name of the property, except for things like typeString.
This caused problems, and I finally wrote TableSchema::allProperties to
resolve.

By the time I was there, I realized the updateDatabase was a mess too.
After much thought, I removed the update piece. We now only ever add new
things. I would be livid if an update over wrote data I had modified to
my liking, so we will not do that to others.

I then figured out why we never updated a postgresql database, and fixed
that. That required some more fixes, because updates come from sqlite. I
had to take care to get the right column names, and to make sure I
handled booleans properly.

Finally, I cleaned a few clangtidy issues up while I was there.

I have tested this many ways. It seems to be solid and stable.