Two small UI adjustments
I swapped the Ca and Cl columns on the water editor. I had intended for
all the cations to be in one column and the anions in the other. The
previous commit got these two backwards.

I also moved the "Brew it!" button into the grid. This way, it stays a
constant size regardless of how big the Style drop down becomes.