First round at fixing BtStringEdit
With the new code to handle high DPI monitors, BtStringEdits were not
being handled correctly. This cascade started with that.

BtStringEdits will now check to see how long their text() is. If the
text() is between 8 and 50 characters (note: unicode characters, not
chars) then the field size gets recalculated. If it is less than 8, we
set it to 8. If it is greated than 50, we only show 50 characters.

This sent me through the code to find all the places that needed
BtStringEdits. Every time I saved the hopEditor, the layouts broke. So
then I reworked the hopEditor. It was such a brilliant idea, I started
fixing all the other editors.

The equipment editor, though, was specified exclusively in code. It
isn't any more, because I cannot edit or do anything without the ui
file. I fixed the Default absorption button too.

And then I learned how to put my tabs on the left, and get the text
flowing in the correct direction.

I am probably half way done with this.