Mostly works
I've been through all the editors. The experience should be fairly
uniform now.

The changes to the ui files are mostly reorganizing the data, a few bug
fixes and getting the tabs in the right order.

Changes to the editors are mostly to support the three buttons. This
will include removing the accept and reject bindings, adding the new*()
methods, making sure tabs are set horizontal and changing the names of
tabs. I've tested what I can think to test on this, and it seems to be
working.

There was an odd typo in TableSchema where PropertyNames::Hop::time_min
was used instead of PropertyNames::Misc::time. That took some digging.
This had a knock on change in misc.cpp

Database::insertElements() needed a similar fix as to what I did PR#564.

There's still an issue in beerxml that I need to chase down on an
import. Once I've figured that out, I will submit the PR.