My half to close #576 - beer xml imports don't seem to be working
This took some work to figure out.

The first issue is anything using allPropertyNames is probably doing it
wrong. Whenever we are storing a type string in the database, there is a
difference in the property names of the objects (which is how we key the
m_properties hash) and the name of the property we store in the
database. This causes many problems and we should fix it some day. Until
then, we should always use allProperties() to loop, and request the
propertyName().

The second issue was a missing property in the
XMLRecord::FieldDefinitions for BEER_XML_MASH_STEP_RECORD_FIELDS. Since
we didn't map the infusion temps to a property, they never got read.

The final issue was that we were treating mashes has having to be
unique. They do not, and enforcing it caused problems when importing two
recipes, each with an unnamed mash.