Closes #580
There are two generateInsert* methods. generateInsertRow() is intended
for database to database copies, as it includes things like foreign
keys. generateInsertProperties is intended for writing entire objects to the
database.

generateInsertRow() needs to be keyed by the column name. We do not have
the object at that point and the object doesn't know the foreign key
names.