For the third time, recipe versioning
I really want this feature. I spend a lot of time writing and tweaking
recipes, and I want an automatic method to version the recipes.

I am basically hand merging from an older branch, because the cherry
picking or rebase is simply trying to cover too much time and too many
changes.

This is the first of about 20 commits. It implements the very basics,
including adding and populating the column if required. It implements
the basics for finding ancestors and displaying them in the Recipe tree.