Added one way to see ancestors
The problem is that the ancestors have display set to false. The filter
will normally not show them. This is one way to over ride that.

Basically, there is an inclusion list that takes precedence over
display. It is set when we ask the tree to show ancestors.

To be honest, I hope this goes away in a later commit. It seems rather
clunky.