The beginnings of an implementation
This commit puts the frame work in place for deciding when a recipe
needs versioning. It also indicates I may have problems with the Undo
system.

The base idea is when anything is added to a recipe, the addToRecipe()
method decides if a new version is required. A new version is required
if the recipe has been brewed at least once right now. More controls
will be added later.

If a version is required, the recipe is spawned. The spawn points to the
ancestor and the whatever caused the spawn will be attached to the
spawn. A signal is then emitted to say a spawn has been created.

MainWindow catches that signal, makes sure the recipe is set to the
spawn and invalidates the filter proxy to force a redraw.

This is a proof of concept working mostly on hops. There's a bunch of
work to be done to make it work every where it needs to work.