Trying to get multi-adds to undo
I tried to do a small change and simply take a single whatToAddOrRemove
and make it a list. That's not work so well, because the method
signatures are all wrong.

If I use two undo classes, there will be two different queues which
won't work.

I am going to have to be more clever about how I handle the single v
multiple use case.