Deleting something from a recipe will version
Deleting is a bit weird now. If versions are being used, deleting an
ingredient actually means "create a new recipe with everything but ..".

This change includes changing the ancestors array from using the
recipe's key() to using the recipe*. It simplifies some things and
doesn't really cost much extra.

This change adds a visual bug to indicate which recipes have ancestors.
The idea is good, but the implementation leaves something to be desired.
I may want to add a "dateForked" to the database so we can display that
information.

I still need to handle:
    - modifying the recipe from the tables and not the trees;
    - allowing a user to lock a recipe and have it version regardless
    - decide if versioning is automatic, or allow the user to turn it
    on/off both globally and per recipe.