Ask Yog-Sothoth about signals
It turns out changing the amount of an ingredient in a recipe is rather
difficult. This is the first step towards making that happen.

The hard part is if I copy recipe A to recipe A' because hop C has
changed in quantity, I have no real way to know which hop to change in
A'. We do not track generational ingredients.

I solve that problem by extending the delete concept. If a recipe wants
versioning when a quantity changes, we copy every but the ingredient
that changed (using spawnWithExclusion). We then copy the changed
ingredient and manually add that to the new recipe. That gives me all
the objects I need to actually change the quantity.

There are all sorts of knock on effects, mostly with how we defined our
signals. Did you know you are not supposed to overload signals? I
didn't, and the only way to make this sane is to overload them and use
templates. However, I think I found a better way to overload them.