Read only
This gets a long way towards being able to mark a recipe as read only.
There are some edge cases to deal with, but the base is in place.

I really need to look at how I am handing the BtTree* classes. I do not
like how they current read.