Signals renamed, database  tables updated
I've renamed all the newSignals to createdSignals. Not sure why I felt
that was required.

I added the code to DatabaseSchemaHelper to also create the locked
column on the recipe table.

I rearranged Datbase::setAncestor to handle all the bits and bobs. It
now sets the ancestor_id, sets locked and display as appropriate. It not
only cleans up the calling code, but it puts everything in a nice
transaction boundary.