Delete now works
What delete means when versioning is a hard question. Does it delete
just the visible recipe, or does it delete all of the ancestors too?

I leave it to the users, and gave an options panel to enable the
versioning and to define what delete means. The default is to not enable
versioning, and to only delete the leaf node.

I cleaned the OptionDialog up a little. We might want to consider
rethinking that UI. It is growing many tabs.