Tweaks and fixes
The recursive query would sometimes return in reverse order. A small
tweak to the query fixes the problem.

The drag/drop code (which will later be removed entirely) gets another
tweak so that dropping recipe A on recipe B will only do the ancestor
thing if A and B are in the same folder.

Included is a nice enhancement to the logo tool tip for sqlite.