Drag and drop no more
As promised, the drag and drop for making ancestors is now gone. It
really is amazingly confusing. So drag and drop for recipes reverts to
its original mode.

This required creating a new dialog and wiring it in properly. If we
need to think about the options dialog, we need to start looking at the
tools menu too.

A few small fixes as I found the issues, but nothing life shattering