Locking on the mainWindow behaves
I had somehow made the lock button global? Anyway, this makes sure the
locked state is set when the recipe is set.

I also moved the brewit button. I wasn't quite a fan of where it was,
and I wanted that space for something else. I still need to find a good
icon for "BrewIt".