ancestor dialog is working
I had the ancestor and the descendant dialog reversed on the form. I
updated MainWindow so it prepopulates the ancestor based on whatever you
have selected in the recipe tree.

I changed "Begat" to "is an ancestor of". While being biblical is fun,
it took me too much thinking.

I had forgotten a commit() in Database::setAncestor().

DatabaseSchemaHelper has to prepopulate the ancestor_id when adding the
columns. The entire system is based on the assumption that the first
ancestor points to itself, and no thing happens if it doesn't.