Bugs, large and subtle
The easy part is I renamed "ancestors" to "snapshots" in the menus. I
hope that makes more sense. I will fix "orphan" just as soon as I can
figure out with what.

The hard parts are three subtle bugs.

The first was that Database::copy simply copied the entire row from the
database -- foreign keys and all. It shouldn't be doing that. Foreign
keys should be left to the calling methods to handle. We didn't see it
because only Recipe has them, and it sort of worked.

The second was that Database::copy then used the source row to create a
new NamedEntity. This was cute, but wrong. I added a new method that
simply fetches an entire row from a table and key, and used that
instead.

The third was that copying a recipe showed it in the wrong folder until
you restarted. If a recipe is in a folder and I copy it, then the copy
is supposed to be in the same folder. It wasn't working that way. Now it
does.

So now copy an ancestor works the way I intended. A copy is created, you
see it immediately, it is in the right folder and it no longer knows who
it's ancestors are. I am not sure that is what I want, but lets start
from here.