I found all the hard problems in programming
Cache. So the core issue was that we updated the cache, then decided to clone the ingredient.

If we decided to clone, then the ancestor's cache had the wrong value because I set the cache before the cloning. Additionally, the clone had the original value cached because of how we copied.

I had to rework the entire caching scheme. I had to return a boolean all the way up the stack from modifyEntry() to indicate if we had cloned or not. It also means we now MUST cache after we clone/set.  Annoying, but so it goes. This means I had to touch basically every model class except BrewNote and Recipe. They never cause a spawn themselves.

I had to create Database::replicate. Database::copy() provided no way to over ride the values and it was all too late by the time T* tmp = new T() was called. So replicate takes a property and value and over rides that property with that value. Just to be difficult, it totally ignore display().

I have declared that inventory changes do not fork recipes. I think that makes sense.

I fixed a few small bugs with things appearing the in the wrong folder again.

I added absoluteZero to the constants file and changed some tests to use it.

I removed a redundant BrewNote::key() as that got implemented in the base class later.

I add the m prefix to the variables in NamedEntity, because the _display broke the macro I wrote to make all the changes required previously. It was obviously faster to change the names than to rework my macro.