Small fixes all over the place. namespace nBtPage is now named BtPage class BtPage is now named Page Enum PlacingFlags for placing of items is split into separate enums, RelationalPlacingFlags and FixedPlacingFlags Include guards hade the leading _, is now renamed with prefix BTPAGE_ intead. private list _children is now named items insted, better name for it anyway. dev docs updated to reflect the changes
One note though, PlacingFlags was a enum struct ... and now I made it to be a ragular enum.
My thinkning here is that this seems to be the norm when creating flags, all of them is accessible through BtPage::<flag> but you
cannot supply the wrong flag to the function due to the declaration. neat!