Fixes Bug in issue #609,
I added a nullcheck for recipe so that when you right click the top-level-item in the recipetree 'Recipes' you get the folder menu but with no DELETE function.

also added the above mentioned folder-specialized context menu.
fixed a few, not all private varables prefixed with _ to be prefixed with m_ instead.