Drag and drop works
The login in BtTreeView::mimeData was broken. As we always set itsa, the
std::optional variable would have never been false. This resulted in
setting the mime type to `application/x-brewtarget-folder` for anything
that wasn't a Recipe, Style or Folder. Since the tables are configured
to only accept `application/x-brewtarget-ingredient`, they wouldn't
accept the drop.

To fix it, I changed all the tests around. We test for Recipe, Style and
Equipment; then we test for Folder; then we test to not be Water. That
sets all of the mime types correct. I also added an
`application/x-brewtarget-water` in case I ever want to get clever.